Complete solution for the Java Stack problem
This is a fun problem and I found a cool way to use stack to see if openings and closings match up in a string since only the most recently one opened matters, and they will be looked at in reverse order as you go up out of the nest.

This would be a good way to check things like methods in a block of code since they open and close with brackets.